refactor(extension): [LW-9094] use new wallet manager
In order to support multi-wallet/account & shared wallets,
walletManager was completely redesigned in SDK.

There are a couple new components:
- WalletRepository stores key agent data, used accounts,
as well as any additional metadata such as 'name'. Lace
should use it to get an array of all users wallets.
- SignerManager is an abstraction over KeyAgent. Lace
now longer needs to maintain the lifecycle of a KeyAgent.
Instead, it should use this component to sign transactions.