fix(extension): connect to hw device asap when adding account
If it takes more than some time (probably 50ms), Chrome will
reject hardware device connection with an error that says
WebHID connections must be initiated from user gesture.

Apparently, fetching wallets from repository can sometimes
take longer. This fix updates useWalletManager addAccount
method to take in the entire AnyBip32Wallet object instead
of just walletId, so that it doesn't have to fetch the
wallet object from indexeddb in service worker.

Consequently, some validations from addAccount can be removed,
because it's signature no longer accepts walletId of a script
wallet. Also it has to trust that wallet actually exists,
otherwise it will reject with an error that comes from the
WalletRepository (it is no longer a responsibility of this
method to check the existence of the wallet).