feat(extension): disable 'Lock wallet' button
Forgot Password flow implementation was deleting the wallet
and opening the Restore Wallet onboarding flow.

When we introduced multi-wallet, deleting the locked/active
wallet left Lace in a faulty state, where it would still have
other wallets remaining, but still locked with the passphrase
of the deleted wallet.

This feature has to be reworked with multi-wallet in mind.
Disabling it for now.