feat: relax tip polling
The SDK default is a 5 second interval, which results in a responsive
outcome but is not appropriate for a centralized remote server. This change
sets a more appropriate interval of 30 seconds, which we may relax even
further if needed. The defaults are provided for `maxInterval` and
`consideredOutOfSyncAfter` since the API doesn't support partial config.