build: separate content scripts build to exclude sentry plugin
Sentry release tracking pollutes global scope by declaring 'e'
variable, which can cause conflicts with dapp code when used
in the injected script.

We were NOT using sentry in the content scripts, so this has
no functional effect. The reason that this code was injected
is because we were using the same webpack build configuration
for both: Lace app scripts and content scripts. Webpack plugins
extend compiler behavior globally (applies to all entrypoints).