fix: navigation to notifications center when last notification is removed
For some reason (at least to me, unknown), the notifications returned by useNotificationsCenter
hook is initially undefined.
This means that when the last notification is removed, the hook returns undefined and the
navigation to the notifications center is not triggered.
Changed the check: when notifications is not undefined, even if an empty array, the navigation
must be triggered.