LW-13832 Integrate NotificationsClient library (#2053)
* feat: integrate NotificationsClient library

* fix: navigation to notifications center when last notification is removed

For some reason (at least to me, unknown), the notifications returned by useNotificationsCenter
hook is initially undefined.
This means that when the last notification is removed, the hook returns undefined and the
navigation to the notifications center is not triggered.
Changed the check: when notifications is not undefined, even if an empty array, the navigation
must be triggered.

* fix: navigation to goBack when a notification is removed

When the condition is met, (either due to error for bad navigation or because the notification is
removed), the correct action is to go back to previous page: pushing a new page injects into the
history a sort of infinite loop-back navigation, as a result the back button does not work.

* fix: isSubscribed in notifications type to reflect the type from @lace/notifications package

The wrong attribute name prevents the toggle switch in the subscriptions dropdown to work.

* feat: strengthen validation of topics and notifications in e2e tests injection script

A typo in an attribute could be a source of waste of time investigating the reason for the not
expected behavior. This gives an immediate feedback in case of error.